ABH-Rush Framework for GameMaker 2023.4+ (PROTOTYPE UPDATE)
None of this is intended to be the finalized version of ABH-Rush for GM 2023.4+
Some of the code is going to be redone and cleaned up alongside the completion of assets for a more clean release.
In the mean time, if you do plan on using it, then be warned that it's incomplete as hell. Support is always accepted.