ABH-Rush Framework for GameMaker 2023.4+ (PROTOTYPE UPDATE 2)
Much like last time, it is not recommended to use it in it's current state, but if you do, then you have been warned.

Following up from last time:
>A new set of HUDs have been added ("Legacy", "Proto", and "Mirai"). While Mirai is planned to be the main default HUD Style for this version of the framework, it is not complete, as it uses the Proto HUD style.
>An improved camera system, while it may not match what the actual GMS1 version does, it does work in it's own way.
> Custom Resolution Commands.
> Starting work on the animation system.

Again, none of this is final, but it should be a good while before things could be at a state where I can call it acceptable for people to use it legitimately.